feat: 优化博客存储

+++ src/main/java/com/isyxf/blog/service/ArticleService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ArticleService {
     void add(Article article);
-    void remove(long id);
+    void remove(int id);
     void edit(Article article);
     Article getById(Integer id);
     List<Article> getAll();
